minor fix migration

+++ database/migrations/2023_09_25_130521_create_apartments_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('address');
             $table->string('longitude')->nullable();
             $table->string('latitude')->nullable();
-            $table->decimal('price', 4, 2)->nullable();
+            $table->integer('price')->nullable();
             $table->smallInteger('beds');
             $table->smallInteger('rooms')->nullable();
             $table->smallInteger('bathrooms')->nullable();
